fix: update unit and spec tests
