+++ tests/test_dockerfile.py
@@ -1,19 +1,8 @@
-import subprocess
+"""Spec tests for container image."""
+
 import time
 
 import pytest
-import testinfra
-
-
-@pytest.fixture(scope="session")
-def host(request):
-    test_image = "docker-centos7-slurm:spec-test"
-    subprocess.check_call(["docker", "build", "-t", test_image, "."])
-    docker_id = subprocess.check_output(
-        ["docker", "run", "-d", "-it", "-h", "ernie", test_image]
-    ).decode().strip()
-    yield testinfra.get_host("docker://" + docker_id)
-    subprocess.check_call(["docker", "rm", "-f", docker_id])
 
 
 @pytest.fixture
@@ -35,7 +24,7 @@ def Slow():
 
 
 def test_tini_is_installed(host):
-    cmd = host.run("tini --version")
+    cmd = host.run("/tini --version")
     assert "0.18.0" in cmd.stdout
 
 
@@ -44,14 +33,10 @@ def test_slurm_user_group_exists(host):
     assert host.user("slurm").group == "slurm"
 
 
-@pytest.mark.parametrize("version, semver", [
-    ("3.6", "3.6.8"),
-    ("3.7", "3.7.5"),
-    ("3.8", "3.8.0")
-])
-def test_python_is_installed(host, version, semver):
+@pytest.mark.parametrize("version", ["3.6", "3.7", "3.8"])
+def test_python_is_installed(host, version):
     cmd = host.run(f"python{version} --version")
-    assert semver in cmd.stdout
+    assert cmd.stdout.startswith(f"Python {version}")
 
 
 def test_mariadb_is_listening(host, Slow):
@@ -78,4 +63,4 @@ def test_slurmctld_is_listening(host, Slow):
 
 def test_slurmd_version(host):
     cmd = host.run("scontrol show config | grep SLURM_VERSION")
-    assert "19.05.4" in cmd.stdout
+    assert "20.02.0" in cmd.stdout
